+++ Docs/ChangeLog.md
@@ -1,3 +1,26 @@
+# Version: 1.119.3
+## Date: 2024-12-3
+### Changes:
+- Fix: Fixed the bug of Logger.
+
+# Version: 1.119.2
+## Date: 2024-12-3
+### Changes:
+- Change: Change the layout of selection group or sample group in the basic plot part to make the layout more clear.
+
+# Version: 1.119.1
+## Date: 2024-11-29
+### Changes:
+- New: Added UpSet plot for basic selection plot part.
+
+
+# Version: 1.119.0
+## Date: 2024-11-28
+### Changes:
+- Change: API updated to 4 
+- New: Added an option to plot the upset plot.
+
+
 # Version: 1.118.4
 ## Date: 2024-11-26
 ### Changes: